Ai-services/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pypdf
chromadb

# LangChain Ecosystem
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
langchain-classic

# Model Providers
langchain-google-genai  
langchain-openai        